Only depend on the fabric API modules you need, fix ModMenu in dev env
The reason ModMenu was causing Mixin to scream in pain is because it depends on a horribly outdated version of Fabric API that doesn't even work in the newest versions of MC. By preventing it from pulling in that dependency, and simply providing a working version yourself, ModMenu will work just fine.

Additionally, there's no need to pull in the entirety of Fabric API if all you need is the key binding API. Likewise, there's no need to declare your dependency on the entirety of Fabric API if all you *really* need is the key binding API.

It's a convenience during development to have the entire API available (as a gradle dependency), but on release you should identify which modules you use and depend on those modules specifically instead of `fabric`. ModMenu does this with fabric-resource-loader-v0, and with this commit RebindNarrator will only require fabric-key-binding-api-v1.